Proje/backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
python-dotenv>=0.21.0
python-multipart>=0.0.5
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
langchain-pinecone>=0.1.0
langchain-community>=0.0.10
pinecone-client>=2.2.0
uuid>=1.30
PyPDF2>=3.0.0
pypdf

